requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
allure-python-commons
requests
faker